Use dark-mode 'class' instead of 'media'
Rationale here is that, in order for CSS vars to work as expected in
components that are used on both docs and marketing pages, the TW dark: syntax
will result in inconsistent styles. The initial setup adds a bit of complexity
to avoid FOUC, but implementing styles everywhere becomes simpler because both
dark: styles and our CSS vars will work more predictably.