Update code element's display property
Changed the display property of the "code" element to "grid". This is because on mobile or smaller screen devices, the length of the "span" element used for highlighting lines of code was not highlighting the entire line of code where it was applied. The issue was caused by the parent element "code" being inline and not filling up the entire width of its parent element, which is the "pre" HTML element.

By setting the display property of the "code" element to "grid", its width now fills the entire width of its parent element "pre". This resolves the issue and makes the "span" element (the green code highlighter) fill the entire width of the code element.

To provide better visualization, I have included before and after screenshots of the issue in my pull request.