Fix code element not assuming full width of parent
This commit resolves the issue of the code element not assuming the full width of its parent element (pre), which caused the highlighting span elements (children of the code element) not to span the entire length of the code they were highlighting.

The issue occurred because the code element is an inline property, which means it flows with surrounding text. However, by setting the display property to inline-block, we can change the display behavior of the element to behave like a block-level element, allowing us to set the width and height properties.

To ensure that the code element assumes the width of its parent container, I changed the display property of code element from grid to inline-block then I  explicitly set the min-width property to 100%. This allows the code element to expand to the full available width of its parent element thereby making the highlighting span element span across the entire line of code regardless of screen size.

Also remove the CSS rule max-width: auto; as it's of no effect.